requirements.txt
google-generativeai
PyPDF2
python-dotenv
sentence-transformers
faiss-cpu
httpx
